Made possible the following experiments over proxy compiled wrf.exe in dm/par configuration configured different dy/dx, restart_interval, history_interval parameters can run ./ideal.exe can run ./wrf.exe in single/multiple process mode (over 5 hr run) can run ./wrf.exe in multiple server mode regressed over iozone, IOR/MPI Fixed the following bugs All the memory-leaking are identified and eliminated constant memory consumption: 10496k virtual memory, with resident 992k(compared with pvfs2-server: 132m virtual memory, 8344k resident) EOF when reading a file is fixed fixing the my_share function by introducing my_offset setting the last_flow bit of the I/O response when encountering EOF

+++ scheduler_SFQD_Full.h
@@ -42,7 +42,7 @@ struct sfqdfull_queue_item
 	int data_socket;
 	int request_port;
 	int request_socket_index;
-	int data_file_size;
+	long long data_file_size;
 	int strip_size;
 	int server_nr;
 	long long file_offset;
@@ -63,9 +63,6 @@ struct sfqdfull_queue_item
 	struct timeval deadline;	//for edf queue
 	struct timeval queuedtime;	//for waiting time calculation
 	struct timeval dispatchtime; // for response time feedback
-
-
-
 };
 
 extern FILE* depthtrack;
